refactor: user and profile serverless functions
